Use f64 for p-values because they often become small.

+++ gwas-summary-statistics-app/src/gwasss.rs
@@ -296,7 +296,7 @@ pub struct AssociationStat {
     pub coded_allele: CodedAllele,
     pub effect: f32,
     pub se: f32,
-    pub p: f32
+    pub p: f64
 }
